# lib/api-zod/src/generated/types/candidate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Candidate {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  currentTitle?: string | null;
  /** @nullable */
  currentCompany?: string | null;
  /** @nullable */
  yearsExperience?: number | null;
  /** @nullable */
  educationLevel?: string | null;
  skills?: string[];
  /** @nullable */
  resumeText?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  activityScore?: number | null;
  createdAt: Date;
}