# lib/api-zod/src/generated/types/candidateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CandidateInput {
  name: string;
  email?: string;
  currentTitle?: string;
  currentCompany?: string;
  yearsExperience?: number;
  educationLevel?: string;
  skills?: string[];
  resumeText?: string;
  location?: string;
  linkedinUrl?: string;
  activityScore?: number;
}