# lib/api-zod/src/generated/types/candidateStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LabelCount } from './labelCount';

export interface CandidateStats {
  total: number;
  avgExperience: number;
  educationBreakdown: LabelCount[];
  topSkills: LabelCount[];
}