# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobInput {
  title: string;
  rawText: string;
}