# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobUpdate {
  title?: string;
  requiredSkills?: string[];
  preferredSkills?: string[];
  minExperience?: number;
  educationRequirement?: string;
  domain?: string;
  seniorityLevel?: string;
}