# lib/api-zod/src/generated/types/labelCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LabelCount {
  label: string;
  count: number;
}