# lib/api-zod/src/generated/types/rankingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RankingInput {
  jobId: number;
  topN?: number;
  minScore?: number;
}