# lib/api-zod/src/generated/types/rankingStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CandidateIQ API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LabelCount } from './labelCount';
import type { ScoreBucket } from './scoreBucket';

export interface RankingStats {
  scoreDistribution: ScoreBucket[];
  avgCompositeScore: number;
  topMatchedSkills: LabelCount[];
  topMissingSkills: LabelCount[];
}